+++ Lib/fontbakery/checks/googlefonts/axisregistry.py
@@ -208,8 +208,9 @@ def com_google_fonts_check_STAT_gf_axisregistry_names(ttFont):
             # Allow an Italic AxisValue for the slnt axis which can be any
             # user value.
             if axis.AxisTag == "slnt" and name == "Italic":
-                yield INFO, Message("italic-detected",
-                                    "Detected an Italic AxisValue for the slnt axis")
+                yield INFO, Message(
+                    "italic-detected", "Detected an Italic AxisValue for the slnt axis"
+                )
                 continue
             if hasattr(axis_value, "Value"):  # Format 1 & 3
                 is_value = axis_value.Value